few changes

+++ backend/src/controllers/trips.ts
@@ -2,8 +2,10 @@ import { RequestHandler } from "express";
 import TripModel from "../models/trip";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getTrips: RequestHandler =  async (req, res, next) => {
+
     try {
         const trips = await TripModel.find().exec();
         res.status(200).json(trips);
@@ -13,10 +15,25 @@ export const getTrips: RequestHandler =  async (req, res, next) => {
     }
 }
 
+export const getMyTrips: RequestHandler =  async (req, res, next) => {
+    const userId = req.session.userId;
+
+    try {
+        assertIsDefined(userId); 
+        const trips = await TripModel.find({userId}).exec();
+        res.status(200).json(trips);
+        
+    } catch (error) {
+        next(error);
+    }
+}
+
 export const getTrip: RequestHandler = async (req, res, next) => {
     const tripId = req.params.tripId;
+
     
     try {
+
         if(!mongoose.isValidObjectId(tripId)) {
             throw createHttpError(400, "invalid trip Id");
         }
@@ -58,13 +75,18 @@ export const createTrip: RequestHandler<unknown, unknown, createTripBody, unknow
     const author = req.body.author;
     const votes = 0;
     const favs = 0;
+    const userId = req.session.userId;
+
 
     try {
+        assertIsDefined(userId); 
+
         if (!title || !body) {
             throw createHttpError(400, "The trip must have a title and a body")
         }
 
         const newTrip = await TripModel.create({
+            userId: userId,
             title: title,
             body: body,
             author: author,
@@ -86,8 +108,12 @@ export const createTrip: RequestHandler<unknown, unknown, createTripBody, unknow
         const newTitle = req.body.title;
         const newBody = req.body.body;
         const newAuthor = req.body.author;
+        const userId = req.session.userId;
+
         
         try {
+            assertIsDefined(userId); 
+
             if(!mongoose.isValidObjectId(tripId)) {
                 throw createHttpError(400, "invalid trip Id");
             }
@@ -100,6 +126,10 @@ export const createTrip: RequestHandler<unknown, unknown, createTripBody, unknow
     
             if(!trip) {
                 throw createHttpError(404, "trip not found");
+            }
+
+            if(!trip.userId.equals(userId)) {
+                throw createHttpError(401, " You naughty naughty, your don't have access to this note.");
             }
     
             trip.title = newTitle;
@@ -115,8 +145,12 @@ export const createTrip: RequestHandler<unknown, unknown, createTripBody, unknow
 
 export const deleteTrip: RequestHandler = async (req, res, next) => {
     const tripId = req.params.tripId;
+    const userId = req.session.userId;
+
 
     try {
+        assertIsDefined(userId); 
+
         if(!mongoose.isValidObjectId(tripId)) {
             throw createHttpError(400, "invalid trip Id");
         }
@@ -125,6 +159,9 @@ export const deleteTrip: RequestHandler = async (req, res, next) => {
 
         if(!trip) {
             throw createHttpError(404, "trip not found");
+        }
+        if(!trip.userId.equals(userId)) {
+            throw createHttpError(401, " You naughty naughty, your don't have access to this note.");
         }
 
         await trip.deleteOne();
